refacctor: Move classes with Framework namespace to Helper
