+++ trunk/src/Helper/class-screen.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Trasweb\Plugins\DisplayMetadata\Framework;
+namespace Trasweb\Plugins\DisplayMetadata\Helper;
 
 use WP_Screen;
